requirements.txt
pywebview>=4.4.1
supabase>=2.0.0
